third_party/blink/web_tests/external/wpt/css/selectors/parsing/parse-has-disallow-nesting-has-inside-has.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] ":has(:is(:has(*)))" should be a valid selector
  assert_equals: serialization should be canonical expected ":has(:is(:has(*)))" but got ":has(:is())"
[FAIL] ":has(:where(:has(*)))" should be a valid selector
  assert_equals: serialization should be canonical expected ":has(:where(:has(*)))" but got ":has(:where())"
Harness: the test ran to completion.
